スクリプト仕分け
