+++ src/Assets/Scenes/Stage/ichi_yama.unity
@@ -15148,8 +15148,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   button: {fileID: 624576008}
-  audioSources: []
-  scheduleOffset: 0.1
 --- !u!1 &625733883
 GameObject:
   m_ObjectHideFlags: 0
