+++ src/Assets/Scripts/Masa script/NoUse/ClearCheck.cs
