+++ src/Assets/Scripts/Masa script/NoUse/Fields/CollectObjectRePop.cs
