+++ src/Assets/Scripts/Masa script/NoUse/GameClearSetiingTest.cs
