+++ src/Assets/Scripts/Masa script/NoUse/UIs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8616ad62bb7b41c4e9589528ee75e57d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
