+++ src/Assets/Scripts/Masa script/NoUse/UIs/DestroyObstcleCount.cs
