+++ src/Assets/Scripts/Masa script/UIs/CountDownSeekBar.cs
@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 using UnityEngine.Events;
-using UnityEngine.SceneManagement;
-using UnityEngine.TextCore.Text;
 
 public class CountDownSeekBar : CountDownTimer
 {
