+++ src/Assets/Scripts/Masa script/UIs/SelectButton.cs
@@ -2,7 +2,6 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
-using System;
 
 public class SelectButton : MonoBehaviour
 {
