+++ src/Assets/Scripts/Masa script/UIs/SpriteAutoTransform.cs
