+++ src/Assets/Scripts/Masa script/UIs/SymbolEmission.cs
@@ -1,6 +1,4 @@
 using System.Collections;
-using System.Collections.Generic;
-using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.UI;
